// frontend/utk-workflow/src/utk-ts/src/interfaces.ts
export interface IExKnot{
    id: string,
    out_name: string;
    in_name?: string;
    group?: IKnotGroup;
    color_map?: string | IConditionBlock;
    range?: number[];
    domain?: number[];
    scale?: string;
}
// ...
export interface IKnot {
    id: string,
    group?: IKnotGroup,
    knot_op?: boolean,
    color_map?: string | IConditionBlock,
    integration_scheme: ILinkDescription[],
    range?: number[],
    domain?: number[],
    scale?: string,
}

export interface IKnotGroup{
    group_name: string,
    position: number
}

export interface ILinkDescription {
    spatial_relation?: SpatialRelationType;
    out: {name: string, level: LevelType};
    in?: {name: string, level: LevelType};
    operation: OperationType | IConditionBlock; 
    abstract?: boolean; // if this is a link with an abstract layer
    op?: string; // TODO: the determination if it is a knot operation or not should be done dinamically
    maxDistance?: number; // max distance for the nearest spatial_relation (meters considering that the CRS of the layers is 3395)
    defaultValue?: number; // default value used in Nearest abstract join in case right side is null
}

/**
 * Interface with the camera definition
 */
export interface ICameraData {
    position: number[]; // camera position and look at
    direction: {right: number[], lookAt: number[], up: number[]} // default: right: 0, 0, 1000. lookAt: 0, 0, 0. up: 0, 1, 0
}

/**
 * Interface with the layer style definition
 */
 export interface IMapStyle {
    land: ColorHEX;     // land layer color definition
    roads: ColorHEX;    // roads layer color definition
    parks: ColorHEX;    // parks layer color definition
    water: ColorHEX;    // water layer color definition
    sky: ColorHEX;      // sky layer color definition
    surface: ColorHEX;
    building: ColorHEX; // buildings layer color definition
}

/**
 * Interface with the layer definition (Feature collection)
 */
export interface ILayerData {
    id: string;                  // layer id
    type: LayerType;             // layer type
    styleKey: keyof IMapStyle;   // layer style key
    data?: ILayerFeature[];      // list of features of the layer 
    renderStyle?: RenderStyle[]; // list of render styles
}
// ...
export interface ILayerFeature {
    // geometries of the feature. (it may have multiple components)
    // geometry: IFeatureGeometry | IFeatureGeometry[];
    geometry: IFeatureGeometry;
    // is highlighted boolean
    highlight?: boolean;
    // what vertices? (if undefined, all vertices are highlighted)
    highlightIds?: number[];
}

/**
 * Interface with the feature geometry definition (Geometry: Geometric info)
 */
export interface IFeatureGeometry {
    coordinates: number[]; // coordinates of the points
    normals?: number[];    // normals over the points
    function?: number[][];   // scalar function defined over the points
    indices?: number[];    // triangles of the geometry
    ids?: number[];    // ids of the cells
    heights?: number[]; // max height of each section of a geometry
    minHeights?: number[]; // min height of each section of a building
    orientedEnvelope?: number[][]; // oriented bounding box around outline of a section of a geometry
    sectionFootprint?: number[][]; // polygon representing the outline of a section of a geometry
    uv?: number[]; // uv value of each coordinate in the surface of a geometry (usually wall)
    width?: number[]; // for each coordinate stores the width of the surface of a geometry (usually wall)
    pointsPerSection?: number[]; // number of points in each section of the geometry
    discardFuncInterval?: number[]; // indicates which coordinates of the geometry to discard in the redering process based on the function values
    varyOpByFunc?: number; // if the opacity should vary according to the function value
}

// The interpreter will replace the block with the end-result of the conditionals
export interface IConditionBlock{
    condition: IConditionElement[],
}

interface IConditionElement{
    test?: string,
    value: any
}